==> LPS/analysis/2-tree.Rmd <==
---
title: "2-tree"
author: 
- name: Ruizhu Huang
  affiliation:
  - &IMLS Institute for Molecular Life Sciences, University of Zurich, Switzerland
date: "2020-04-20"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

# Load packages
```{r}
suppressPackageStartupMessages({
  library(Seurat)
  library(SingleCellExperiment)
  library(clustree)
  library(dplyr)
  library(ggtree)
  library(ggnewscale)
  library(cowplot)
  library(TreeSummarizedExperiment)
  library(TreeHeatmap)
})
```

# load data

```{r}
ao <- readRDS(file.path("output", "ao.rds"))
as <- readRDS(file.path("output", "as.rds"))
```


```{r}
ao_list <- lapply(split(colnames(ao), ao$sample_id), FUN=function(x) subset(ao, cells=x))
```

# Type markers
## Cannonical markers
```{r}
known_markers <- list(
    astrocytes = c("Aqp4", "Gfap", "Fgfr3"),
    endothelial = c("Cldn5","Nostrin"),
    microglia = c("C1qb","Tyrobp"),
    neuronal = c("Snap25", "Stmn2"),
    neuronal_excitatory = "Slc17a7",
    neuronal_inhibitory = "Gad1",
    oligodendrocyte = "Opalin",
    OPC = "Pdgfra")
known_mk <- match(unlist(known_markers), 
                  gsub(".*\\.", "", rownames(ao@assays$RNA)))
known_mk <- rownames(ao@assays$RNA)[known_mk]


```


## Computationally-identified markers

For each cell type, we select 20 markers that have the highest absolute value of log fold change. This is performed separately for samples in control group (`WT`) to avoid selecting state-marker genes.
```{r}
sample_WT <- ao@meta.data %>% 
  filter(group_id == "WT") %>%
  select(sample_id) %>%
  unlist() %>%
  unique()

mk_list <- lapply(ao_list[sample_WT], FUN = function(x){
  FindAllMarkers(x, logfc.threshold = 0.5,
                     verbose = FALSE)})
saveRDS(mk_list, file = "temp/mk_list.rds")
```


```{r}
# Find top N markers
topN <- 20

sel_list <- lapply(mk_list, FUN = function(x) {
  x %>%
    group_by(cluster) %>%
    top_n(topN, abs(avg_logFC)) %>%
    ungroup(cluster) %>%
    select(gene) %>%
    unlist() %>%
    unique()})

# select markers in at least 2 samples (50%)
sel_u <- unlist(sel_list)
length(unique(sel_u))
tb <- table(sel_u)
sel <- names(tb[tb >= 2])
length(sel)

# type markers: known markers & selected markers
sum(known_mk %in% sel)

length(setdiff(known_mk, sel))
all_mk <- unique(c(sel, known_mk))
length(all_mk)
saveRDS(all_mk, file = "temp/all_mk.rds")

# integrate data (using type markers + default(anchor.feature))
ft <- unique(c(as@anchor.features, all_mk))
ias <- IntegrateData(anchorset = as, 
                    features.to.integrate = ft,
                    dims = 1:20, verbose = FALSE)

length(ias@assays$integrated@var.features)
dim(ias@assays$integrated@data)
ias@assays$integrated@scale.data

# scale data (using type markers)
ias <- ScaleData(ias, features = ft, verbose = FALSE)
length(ias@assays$integrated@var.features)
dim(ias@assays$integrated@scale.data)
dim(ias@assays$integrated@data)
```

Cells are clustered using all type markers.
```{r}
# run PCA using type markers
ias <- RunPCA(ias, features = all_mk, 
             npcs = 30, verbose = FALSE)
saveRDS(ias, file = "temp/ias.rds")

# t-SNE and Clustering
ias <- RunUMAP(ias, reduction = "pca", dims = 1:20, 
               verbose = FALSE)
ias <- FindNeighbors(ias, 
                    reduction = "pca", dims = 1:20, 
                    verbose = FALSE)
```


# Tree construction
Leaves of the tree are clusters generated specified resolution.
```{r}
reso <- 6
ias <- FindClusters(ias, resolution = reso, verbose = FALSE)
ias <- BuildClusterTree(object = ias,
                        dims = 1:20)
cell_tree <- Tool(object = ias, slot = 'BuildClusterTree')
cell_tree
#saveRDS(cell_tree, file = "output/cell_tree.rds")
```


# Viz tree
Within cluster, cell counts in each cell type are calcuated.

```{r}
# information: cell types + cluster ID
reso_vn <- paste0("integrated_snn_res.", reso)
cell_info <- ao@meta.data %>% 
  select("sample_id", "barcode", "group_id", "cell_type") %>% 
  mutate(cell = rownames(ao@meta.data))  %>% 
  left_join(ias@meta.data[, c("sample_id", "barcode", "group_id",
                      reso_vn)]) %>%
  rename(cluster_id = !!reso_vn)

# cell type of each leaf on the tree
aa <- table(cell_info$cell_type, cell_info$cluster_id)
ind <- apply(aa, 2, which.max)
cell_anno <- data.frame(cluster_id = colnames(aa),
                        node = transNode(tree = cell_tree, 
                                         node = colnames(aa)),
                        cell_type = rownames(aa)[ind])
rownames(cell_anno) <- cell_anno$cluster_id

```

Leaves of the tree are annotated with cell types. The bar plot displays cell counts of cell types within each cluster. The cell type that has maximum counts in a cluster is used to annotate the leaf.
```{r}

# Tree
Fig_0 <- ggtree(cell_tree, ladderize = TRUE)

# Tree + cell type
d <- dplyr::filter(Fig_0$data, isTip) %>% select(c(label, y)) 

dd <- cell_info %>%
  rename(label = cluster_id) %>%
  left_join(d, by='label')


colr <- c("Astrocytes" = "#E41A1C", "CPE cells" = "#377EB8",
          "Endothelial" = "#4DAF4A", "Excit. Neuron" = "#984EA3",
          "Inhib. Neuron" = "#FFFF33", "Microglia" = "#FF7F00",
           "Oligodendrocytes" = "#A65628", "OPC" = "#F781BF")
bar <- ggplot(dd, aes(x = y)) +
    geom_bar(aes(fill = cell_type)) +
    scale_fill_manual(values = colr) +
    theme(legend.position = "none") +
    theme(axis.line.x=element_blank(),
          axis.text.y=element_blank(),
          axis.ticks.y=element_blank(),
          axis.title.x=element_blank(),
          axis.title.y=element_blank(),
          legend.position="none",
          panel.background=element_blank(),
          panel.border=element_blank(),
          panel.grid.major=element_blank(),
          panel.grid.minor=element_blank(),
          plot.background=element_blank())


bar_df <- ggplot_build(bar)$data[[1]]
vf <- bar_df$fill
names(vf) <- vf

# flip x-y coordinates and scale x 
scale_bar <- 1/300
ux <- max(bar_df$ymax)
lx <- min(bar_df$ymin)

# barplot: gap to the tree (2)
gap_tree <- ceiling(max(Fig_0$data$x)) + 0.1

#
ct <- names(colr)
names(ct) <- colr
dat_bar <- bar_df %>%
  select(fill, x, xmin, xmax, ymin, ymax) %>%
  rename(ymin = xmin,
         ymax = xmax, 
         xmin = ymin, 
         xmax = ymax,
         y = x) %>%
  mutate(xmin_s = (xmin-lx)/(ux-lx)*ux*scale_bar+ gap_tree,
         xmax_s = (xmax-lx)/(ux-lx)*ux*scale_bar+ gap_tree,
         Cell_type = factor(ct[fill], levels = rev(ct)) )
  
   
# label cell counts
x_bar_label <- c(0, 500, 1000, 1500)
names(x_bar_label) <- as.character(x_bar_label*scale_bar + gap_tree)

dat_bar_ext <- dat_bar %>%
  mutate(xmin = min(xmin_s), xmax = max(xmax_s))


Fig_1 <- Fig_0 + 
  geom_segment(data = dat_bar_ext, 
               aes(x = xmin, xend = xmax,
                   y = y, yend = y), 
               linetype = "solid",
               color = "grey80", size = 0.05, 
               show.legend = FALSE)+
  geom_rect(data = dat_bar,
            aes(xmin = xmin_s, ymin = ymin, 
                xmax = xmax_s, ymax = ymax, 
                fill = Cell_type), 
            inherit.aes = FALSE) +
  scale_fill_manual(values = colr) +
  geom_segment(data = dat_bar,
               aes (x = min(xmin_s), y = -0.5, 
                    xend = max(xmax_s), yend = -0.5),
               size = 0.4) +
  annotate("segment", 
           x = as.numeric(names(x_bar_label)), 
           y = rep(-0.2, length(x_bar_label)),
           xend = as.numeric(names(x_bar_label)), 
           yend = rep(-0.8, length(x_bar_label)),
           size = 0.5) +
  annotate("text", 
           x = as.numeric(names(x_bar_label)), 
           y = rep(-1.3, length(x_bar_label)),
           label= x_bar_label, size = 1.5) +
  annotate("text", x = mean(min(dat_bar$xmin_s),max(dat_bar$xmax_s)),
           y = max(dat_bar$ymax) + 3, label = "Cell counts", 
           hjust =0, fontface = "bold", size = 3) +
  guides(fill = guide_legend(order = 2, ncol = 2)) +
  new_scale_fill() +
  theme(legend.position = c(0.1, 0.6))


Fig_1 

  
```

# Output SE

## Type genes

The data of type genes is saved for later visualization as a `SingleCellExperiment` object. 

```{r}
# scale data of cell type marker
mkc <- ias@assays$integrated@scale.data[all_mk, ]

# cell information
cell_dt <- cbind.data.frame(
    cell = rownames(ias@meta.data),
    ias@meta.data[, c("sample_id", "barcode", "group_id",
                      reso_vn)]) %>%
    rename(cluster_id = !!reso_vn) %>% 
    left_join(cell_anno) 


se_mk <- SummarizedExperiment(assays = list(mkc), 
                               colData = cell_dt)
#saveRDS(se_mk, file = "output/se_mk.rds")
```


## Genes without type genes

The original data of genes that are not selected as type genes for all cells is output as a
`SummarizedExperiment` object. The data is later aggregated the different
levels of the tree to compare between groups `LPS` and `WT`.

```{r}
(test <- all(rownames(ao@meta.data) == rownames(ias@meta.data)))
if (!test) {stop("the order of genes are changed")}

# the original data
DefaultAssay(ias) <- "RNA"
count <- ias@assays$RNA@counts
count <- count[!rownames(count) %in% all_mk, ]

cell_info <- ao@meta.data %>% 
  select("sample_id", "barcode", "group_id", "cell_type") %>% 
  mutate(cell = rownames(ao@meta.data))  %>% 
  left_join(ias@meta.data[, c("sample_id", "barcode", "group_id",
                      reso_vn)]) %>%
  rename(cluster_id = !!reso_vn) 



se_count <- SummarizedExperiment(assays = list(count),
                                  colData = cell_info)

```



```{r}
saveRDS(known_mk, file = "output/known_mk.rds")
saveRDS(cell_tree, file = "output/cell_tree.rds")
saveRDS(se_mk, file = "output/se_mk.rds")
saveRDS(se_count, file = "output/se_count.rds")
saveRDS(cell_info, file = "output/cell_info.rds")
saveRDS(all_mk, file = "output/all_mk.rds")
```
