==> LPS/analysis/4-visualization.Rmd <==
---
title: "4-visualization"
author: "fionarhuang"
date: "2020-04-20"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE, cache=FALSE}
knitr::opts_knit$set(echo = TRUE, message = FALSE, warning = FALSE, error=TRUE)
```

# Load packages
```{r}
suppressPackageStartupMessages({
  library(treeclimbR)
  library(TreeSummarizedExperiment)
  library(parallel)
  library(TreeHeatmap)
  library(dplyr)
  library(ggplot2)
  library(ggtree)
  library(ggnewscale)
  library(viridis)
  library(ggrepel)
  library(scales)
  library(cowplot)
  library(ggplotify)
})
```

# load data

```{r}
known_mk <- readRDS("output/known_mk.rds")
se_mk <- readRDS("output/se_mk.rds")
load("output/3-treeclimbR.RData")
cell_info <- readRDS("output/cell_info.rds")
colData(d_se)$group_id <- ifelse(colData(d_se)$group_id == "WT", "Vehicle", "LPS")

# tree
cell_tree <- colTree(tse)
leaf <- showNode(tree = cell_tree, only.leaf = TRUE)
```

# Display Tree
```{r}
# Tree
Fig_0 <- ggtree(cell_tree, ladderize = TRUE)
Fig_0
```


# Display cell counts

```{r}
d <- dplyr::filter(Fig_0$data, isTip) %>% select(c(label, y)) 

dd <- cell_info %>%
  rename(label = cluster_id) %>%
  left_join(d, by='label')


colr <- c("Astrocytes" = "#E41A1C", "CPE cells" = "#377EB8",
          "Endothelial" = "#4DAF4A", "Excit. Neuron" = "#984EA3",
          "Inhib. Neuron" = "#FFFF33", "Microglia" = "#FF7F00",
           "Oligodendrocytes" = "#A65628", "OPC" = "#F781BF")
bar <- ggplot(dd, aes(x = y)) +
    geom_bar(aes(fill = cell_type)) +
    scale_fill_manual(values = colr) +
    theme(legend.position = "none") +
    theme(axis.line.x=element_blank(),
          axis.text.y=element_blank(),
          axis.ticks.y=element_blank(),
          axis.title.x=element_blank(),
          axis.title.y=element_blank(),
          legend.position="none",
          panel.background=element_blank(),
          panel.border=element_blank(),
          panel.grid.major=element_blank(),
          panel.grid.minor=element_blank(),
          plot.background=element_blank())


bar_df <- ggplot_build(bar)$data[[1]]
vf <- bar_df$fill
names(vf) <- vf

# flip x-y coordinates and scale x 
scale_bar <- 1/300
ux <- max(bar_df$ymax)
lx <- min(bar_df$ymin)

# barplot: gap to the tree (2)
gap_tree <- ceiling(max(Fig_0$data$x)) + 0.1

#
ct <- names(colr)
names(ct) <- colr
dat_bar <- bar_df %>%
  select(fill, x, xmin, xmax, ymin, ymax) %>%
  rename(ymin = xmin,
         ymax = xmax, 
         xmin = ymin, 
         xmax = ymax,
         y = x) %>%
  mutate(xmin_s = (xmin-lx)/(ux-lx)*ux*scale_bar+ gap_tree,
         xmax_s = (xmax-lx)/(ux-lx)*ux*scale_bar+ gap_tree,
         Cell_type = factor(ct[fill], levels = rev(ct)) )
  
   
# label cell counts
x_bar_label <- c(0, 500, 1000)
names(x_bar_label) <- as.character(x_bar_label*scale_bar + gap_tree)

dat_bar_ext <- dat_bar %>%
  mutate(xmin = min(xmin_s), xmax = max(xmax_s))


Fig_1 <- Fig_0 + 
  geom_segment(data = dat_bar_ext, 
               aes(x = xmin, xend = xmax,
                   y = y, yend = y), 
               linetype = "solid",
               color = "grey80", size = 0.05, 
               show.legend = FALSE)+
  geom_rect(data = dat_bar,
            aes(xmin = xmin_s, ymin = ymin, 
                xmax = xmax_s, ymax = ymax, 
                fill = Cell_type), 
            inherit.aes = FALSE) +
  scale_fill_manual(values = colr) +
  geom_segment(data = dat_bar,
               aes (x = min(xmin_s), y = -0.5, 
                    xend = max(xmax_s), yend = -0.5),
               size = 0.4) +
  annotate("segment", 
           x = as.numeric(names(x_bar_label)), 
           y = rep(-0.2, length(x_bar_label)),
           xend = as.numeric(names(x_bar_label)), 
           yend = rep(-0.8, length(x_bar_label)),
           size = 0.5) +
  annotate("text", 
           x = as.numeric(names(x_bar_label)), 
           y = rep(-1.3, length(x_bar_label)),
           label= x_bar_label, size = 1.5) +
  annotate("text", x = mean(min(dat_bar$xmin_s),max(dat_bar$xmax_s)),
           y = max(dat_bar$ymax) + 3, label = "Cell counts", 
           hjust =0, fontface = "bold", size = 2) +
  guides(fill = guide_legend(order = 2, ncol = 2)) +
  new_scale_fill() +
  theme(legend.position = c(0.1, 0.6))


Fig_1 
```

# Dispaly type markers

```{r}
## median value of known-marker genes in each leaf
med_mk <- medianByClusterMarker(SE = se_mk, 
                                marker_in_column = FALSE, 
                                column_cluster = "cluster_id", 
                                use_marker = NULL)
mk <- med_mk[, colnames(med_mk) %in% known_mk]
mat_mk <- assays(mk)[[1]]
colnames(mat_mk) <- sub(pattern = ".*?\\.", 
                        replacement = "", colnames(mat_mk))
mat_mk <- mat_mk[, sub(pattern = ".*?\\.", 
                       replacement = "", known_mk)]

# column normalized
mat_type <- apply(mat_mk, 2, FUN = function(x) {
    xx <- (x - min(x))/(max(x) - min(x))
    return(xx)
})
rownames(mat_type) <- rownames(mat_mk)
colnames(mat_type) <- colnames(mat_mk)


# Tree + cell type + (cell type marker gene)
Fig_2 <- TreeHeatmap(tree = cell_tree, tree_hm_gap = 4.8, 
                     rel_width = 0.6,
                     tree_fig = Fig_1, hm_data = mat_type,
                     title_hm = "Genes (type)", show_title = TRUE, 
                     title_offset_y = 8, title_size = 3,
                     show_colnames = TRUE, colnames_offset_y = 2.5,
                     colnames_angle = 90, colnames_size = 1.8,
                     show_rownames = FALSE, 
                     rownames_offset_x = 0.5) +
    scale_fill_viridis_c(guide = "colourbar") + 
    guides(fill = guide_legend(order = 3)) +
    new_scale_fill()
Fig_2

```


# Display state markers
```{r}
## state-marker genes
out_sel <- topNodes(object = best_0.05, n = Inf, p_value = 0.05) 
head(out_sel)
out_sel <- out_sel %>%
  filter(abs(logFC) >= 1) 

list_sel <- split(out_sel, f = out_sel$feature)
gene_node <- lapply(list_sel, FUN = function(x) {
    x$node
})
length(gene_node)

sign_sel <- lapply(list_sel, FUN = function(x) {
    sign(x$logFC)
})
gene <- sub(pattern = ".*?\\.", replacement = "",
            names(gene_node))

df_gene <- data.frame(gene_id = names(gene_node),
                      gene_sid = gene)

head(df_gene)

mat_state <- lapply(seq_along(gene_node), FUN = function(x) {
    node.x <- gene_node[[x]]
    leaf.x <- findOS(tree = cell_tree, node = node.x, 
                     only.leaf = TRUE, self.include = TRUE)
    ll.x <- lapply(leaf.x, length)
    
    sign.x <- sign_sel[[x]]
    sl.x <- rep(sign.x, unlist(ll.x))
    
    vv <- rep(0, length(leaf))
    ind <- match(unlist(leaf.x), leaf)
    vv[ind] <- sl.x
    return(vv)})
mat_state <- do.call(cbind, mat_state)    
colnames(mat_state) <- gene
rownames(mat_state) <- transNode(tree = cell_tree, 
                                node = leaf, use.alias = TRUE)


# Tree + cell type + (cell state gene)
Fig_3 <- TreeHeatmap(tree = cell_tree, tree_fig = Fig_2, 
                     rel_width = 2, tree_hm_gap = 13.8,
                     hm_data = mat_state, cluster_column = TRUE,
                     show_rownames = FALSE, rownames_offset_x = 0.5,
                     rownames_size = 3, rownames_hjust = 0,
                     legend_title_hm = "Change",
                     title_hm = "Genes (state)", show_title = TRUE, 
                     title_offset_y = 4, title_size = 3, 
                     #dist_method = "manhattan", 
                     dist_method = "binary", 
                     hclust_method = "median") +
    scale_fill_gradientn(colors = c("blue", "grey", "orange"), 
                         breaks = c(1, 0, -1),
                         labels = c("Up", "Same", "Down"),
                         guide = guide_legend(order = 3)) + 
    guides(fill = guide_legend(order = 4, ncol = 2)) +
    theme(
      plot.margin = margin(t = 0, b = 0, r = 1, l = 1, "mm"),
      legend.margin = margin(-1,0,0,0, unit="mm"),
      legend.position = c(0.1, 0.65),
      legend.background = element_rect(fill = NA),
      legend.key.size = unit(0.3,"cm"),
      legend.text = element_text(size = 6),
      legend.title = element_text(size = 7),
      legend.spacing.x = unit(1, "mm"),
      legend.spacing.y = unit(2, "mm")
      ) +
  ylim(c(-2, 70))
    
Fig_3

```




# Annotate categories

Below figure is to get node labels for cell types
```{r}
TreeHeatmap(tree = cell_tree, tree_fig = Fig_2, 
                     rel_width = 1.5, tree_hm_gap = 13.8,
                     hm_data = mat_state, cluster_column = TRUE,
                     show_rownames = TRUE, rownames_offset_x = 0.5,
                     rownames_size = 3, rownames_hjust = 0,
                     legend_title_hm = "Change",
                     title_hm = "Genes (state)", show_title = TRUE, 
                     title_offset_y = 4, title_size = 4, 
                     #dist_method = "manhattan", 
                     dist_method = "binary", 
                     hclust_method = "average") +
    scale_fill_gradientn(colors = c("blue", "grey", "orange"), 
                         breaks = c(1, 0, -1),
                         labels = c("Up", "Same", "Down"),
                         guide = guide_legend(order = 3)) + 
    guides(fill = guide_legend(order = 2, ncol = 2)) +
    theme(
      plot.margin = margin(t = 0, b = 0, r = 1, l = 1, "mm"),
      legend.margin = margin(-1,0,0,0, unit="mm"),
      legend.position = c(0.1, 0.65),
      legend.background = element_rect(fill = NA),
      legend.key.size = unit(0.25,"cm"),
      legend.text = element_text(size = 6),
      legend.title = element_text(size = 7),
      legend.spacing.x = unit(1, "mm"),
      legend.spacing.y = unit(2, "mm")
      ) +
  ylim(c(-2, 70)) 
```



```{r}
# cell-type specific branch
# node number: remove prefix "alias_"
root <- setdiff(cell_tree$edge[, 1], cell_tree$edge[, 2])
micro_node <- 48
olig_node <- shareNode(tree = cell_tree, node = c(16, 31, 34))
endo_node <- shareNode(tree = cell_tree, node = c(43, 54))
#glial_node <- shareNode(tree = cell_tree, node = c(63, 1, 39, 53, 46))
astr_node <- shareNode(tree = cell_tree, node = c(41, 44, 23, 24))


node_list <- list(A = root,  B = endo_node, C = astr_node, D = olig_node, 
                  #E = glial_node,  
                  E = micro_node) 
br_list <- findOS(tree = cell_tree, node = unlist(node_list),
                  only.leaf = FALSE, self.include = TRUE)
names(br_list) <- names(node_list)

br_list0 <- br_list
#br_list0$D <- br_list0$C
leaf_list <- findOS(tree = cell_tree, node = unlist(node_list),
                  only.leaf = TRUE, self.include = TRUE)
list_fv <- c(1, as.list(rep(0.4, 4)))
list_l <- list(A = 0.8, B = 0.5, C = 0.75, D = 0.5, E = 1)
list_u <- list(1, 1, 1, 1, 1)
list_nseq <- list(30, 100, 50, 30, 30)
names(list_fv) <- names(list_l) <- names(list_u) <-names(list_nseq) <- names(node_list)

# Data of last heatmap
ds_df <- getData(tree_hm = Fig_3, type = "heatmap")
is_seq <- function(x, by, num_seq, tol = 1e-5){
  x1 <- c(FALSE, abs(diff(x) - by) > tol)
  x2 <- cumsum(x1)+1
  x3 <- table(x2)
  x4 <- which(x3 >= num_seq)
  fx <- x[x2 %in% x4]
  return(fx)
}

list_df <- list_range <- vector("list", length(node_list))
for (i in seq_along(node_list)) {
  x_f <- list_fv[[i]]
  x_br <- br_list[[i]]
  x_br0 <- br_list0[[i]]
  x_leaf <- leaf_list[[i]]
  x_ru <- list_u[[i]]
  x_rl <- list_l[[i]]
  x_nseq <- list_nseq[[i]]
  
  x_df <- ds_df %>%
    group_by(variable) %>%
    mutate(value_sum = sum(abs(value))) %>%
    filter(value_sum <= x_f*length(leaf)) %>% # avoid DS in all cells are picked in cell-type specific
    filter(node %in% x_br0) %>%
    mutate(value_sum = sum(abs(value))) %>% # recalculate the sum of value  
    filter(value_sum <= x_ru*length(x_leaf) & 
             value_sum >= x_rl*length(x_leaf)) %>% 
    filter(node %in% x_br) %>% 
    filter(abs(value) == 1) %>% 
    dplyr::select(variable, x, value_sum) %>%
    distinct() %>%
    arrange(x) 
   
  width <- sort(unique(abs(unique(diff(ds_df$x)))))[2]
  x_seq <- is_seq(x = x_df$x, by = width,
                  num_seq = x_nseq)
  x_df <- x_df %>%
    filter(x %in% x_seq) %>%
    dplyr::select(variable, x) %>%
    distinct() %>%
    arrange(x)
  
  list_df[[i]] <- x_df
  list_range[[i]] <- cbind.data.frame(xmin = range(x_df$x)[1],
                        xmax = range(x_df$x)[2],
                        y = min(ds_df$y) - 2,
                        label= names(node_list)[i],
                        num = nrow(x_df),
                        stringsAsFactors = FALSE)
}

df_range <- do.call(rbind, list_range) 

range_lab <- df_range %>%
  group_by(label) %>%
  summarize(y = unique(y),
            n = sum(num),
            x = mean(range(c(xmin, xmax))))



range_col <- c("#FDD262", "#D3DDDC", "#C7B19C",
               "#D69C4E", "#ABDDDE", "#74A089" )
names(range_col) <- unique(df_range$label)
Fig_4 <- Fig_3 +
  new_scale_color() +
  geom_segment(data = df_range,
               aes(x =xmin, xend = xmax,
                   y = y-0.25, yend =y-0.25, color = label),
               size = 5, show.legend = FALSE,
               inherit.aes = FALSE) +
  scale_colour_manual(values = range_col) +
  geom_text(data = range_lab, 
            aes(x, y, label = label),
            size = 4) +
  theme(
      plot.margin = margin(t = 0, b = 0, r = 1, l = 1, "mm"),
      legend.margin = margin(-1,0,0,0, unit="mm"),
      legend.position = c(0.1, 0.55),
      legend.background = element_rect(fill = NA),
      legend.key.size = unit(0.25,"cm"),
      legend.text = element_text(size = 6),
      legend.title = element_text(size = 7),
      legend.spacing.x = unit(1, "mm"),
      legend.spacing.y = unit(3, "mm")
      ) +
  ylim(c(-2, 70))


Fig_4
```


# DS: volcano plot

```{r}
gene_10 <- node_list
gene_30 <- node_list
dat_vcn <- list_df

for (i in seq_along(node_list)) {
  x_df <- list_df[[i]]
  x_node <- node_list[[i]]
  
  v_df <- x_df$variable
  v_gene <- df_gene %>%
    filter(gene_sid %in% v_df) %>%
    mutate_if(is.factor, as.character)
  
  x_out <- out %>%
    filter(node == x_node) %>%
    filter(feature %in% v_gene$gene_id) %>%
    mutate(gene_sid = sub(pattern = ".*?\\.", replacement = "", feature)) 
  table(sign(x_out$logFC))
  
  t_30  <- sort(abs(x_out$logFC), decreasing = TRUE)[1:30]
  dat_vcn[[i]]  <- x_out %>%
    mutate(top_30 = ifelse(abs(logFC) %in% t_30, "Yes", "No"),
           label = names(node_list)[i],
           num_up = table(sign(x_out$logFC))["1"],
           num_down = table(sign(x_out$logFC))["-1"],
           ngene = nrow(x_out))
  gene_30[[i]] <- dat_vcn[[i]] %>%
    filter(abs(logFC) %in% t_30) %>%
    arrange(logFC) %>%
    select(feature) %>% unlist() %>% as.character()
}
dat_vcn <- do.call(rbind, dat_vcn) 
#%>%  mutate(gene_top30 = ifelse(top_30 == "Yes",gene_sid, ""))

head(dat_vcn)

fig_vcn <- ggplot(dat_vcn, aes(x = logFC, y = -log10(PValue))) +
  geom_vline(xintercept = 0, linetype = "dashed") +
  annotate("text",
            x = -5.3, y = 45, label = "gene:",
            hjust = 0, size = 2) +
  geom_text(data = . %>% select(num_down, label) %>% distinct,
            x = -1.8, y = 45, aes(label = paste(num_down)),
            hjust = 0, size = 2) +
  geom_text(data = . %>% select(num_up, label) %>% distinct,
            x = 0.8, y = 45, aes(label = paste0(num_up)),
            hjust = 0, size = 2) +
  facet_wrap(~label, nrow = 1) +
  geom_point(aes(color = top_30), size = 0.5) +
  scale_color_manual(values = c("black", "red")) +
  guides(color = guide_legend(override.aes = list(size = 2))) +
  theme_bw(base_size = 8) + 
  theme(
    aspect.ratio = 1,
    #plot.margin = unit(c(0, 0, 0.2, 0), "cm"),
    #panel.grid = element_blank(),
    panel.spacing = unit(0, "lines"),
    axis.text = element_text(color = "black"),
    legend.key.size= unit(2, "mm"),
    plot.title = element_text(hjust = 0.5),
    legend.text = element_text(size = 6),
    legend.title = element_text(size = 7),
    legend.position="right",
    legend.margin = margin(t = 0, b = 0, r = 2, l = 0),
    legend.box.margin=margin(-5, -5, -10, -5),
    strip.background = element_rect(colour = "black", fill = "gray90"),
    strip.text.x = element_text(color = "black", size = 10),
    strip.text.y = element_text(color = "black", size = 10)) +
    ylab(expression(-log[10](p)))
fig_vcn

# change the fill color of panel title
g <- ggplot_gtable(ggplot_build(fig_vcn))
stripr <- which(grepl('strip-t', g$layout$name))
fills <- range_col
k <- 1
for (i in stripr) {
  j <- which(grepl('rect', g$grobs[[i]]$grobs[[1]]$childrenOrder))
  g$grobs[[i]]$grobs[[1]]$children[[j]]$gp$fill <- fills[k]
  k <- k+1
}
fig_vcn <- as.ggplot(g)
fig_vcn
```



# top-30: heatmap

```{r}

alias_list <- lapply(node_list, FUN = function(x){
  transNode(tree = cell_tree, node = x, use.alias = TRUE)
})
fig_hp <- vector("list", length(alias_list))

for (i in seq_along(alias_list)) {
  x_alias <- alias_list[[i]]
  x_gene <- gene_30[[i]]

# observed counts in samples
mat_1 <- assays(d_se)[[x_alias]]
mat_1 <- edgeR::cpm(mat_1)

mat_2 <- mat_1[x_gene, ]
mat_3 <- t(apply(mat_2, 1, scale)) 
colnames(mat_3) <- colnames(mat_2)
rownames(mat_3) <- sub(pattern = ".*?\\.", replacement = "", x_gene)

x_tree <- ape::as.phylo(hclust(dist(mat_3)))
x_fig <- ggtree(x_tree)
col_split <- colData(d_se)$group_id
names(col_split) <- colnames(d_se)

split_label <- col_split
names(split_label) <- split_label
fig.i <- Heatmap(hm_data = mat_3, 
            column_split = col_split, 
            column_split_label = split_label,
            column_split_gap = 0.05, 
            column_split_label_size = 2, 
            column_split_label_offset_y = 1.5,
            show_row_tree = FALSE, 
            show_rownames = TRUE,
            rownames_size = 2,
            rownames_position = "left",
            rownames_offset_x = -0.01, 
            rownames_color = "red",
            rownames_hjust = 1, 
            legend_title_hm = "Expression",
            cluster_row = FALSE
            ) +
  expand_limits(x = -10) +
  scale_fill_viridis(option = "D", 
                     limits = c(-1, 1),
                     oob = squish)
dat.i <- getData(tree_hm = fig.i, type = "row_name")
lab.i <- dat.i %>%
  mutate(y = max(y), x = min(x), 
         label = names(alias_list)[i]) %>%
  select(x, y, label) %>%
  distinct()
fig_hp[[i]] <- fig.i + 
  geom_rect(data = lab.i, 
            aes(xmin = x-4, xmax = x-0.4, 
                ymin = y + 0.5, ymax = y+2), 
            fill = range_col[lab.i$label]) +
  annotate("text", x = lab.i$x - 2.2, 
           y = lab.i$y + 1.25, label = lab.i$label)
}

legend <- get_legend(
  fig_hp[[1]] + 
    theme(legend.box.margin = margin(t = 0, b = 0, r = 5, l = -5),
          legend.key.size= unit(3, "mm"),
          legend.title = element_text(size = 7),
          legend.text = element_text(size = 6))
)
fig_b1 <- plot_grid(
          fig_hp[[1]] + theme(plot.margin = margin(t = -5, b = 0, 
                                       r = 1, l = 0, "mm"), 
                              legend.position="none"),
          fig_hp[[2]] + theme(plot.margin = margin(t = -5, b = 0,
                                                   r = 1, l = -8, "mm"),
                              legend.position="none"),
          fig_hp[[3]] + theme(plot.margin = margin(t = -5, b = 0, 
                                                   r = 1, l = -9, "mm"),
                              legend.position="none"),
          fig_hp[[4]] + theme(plot.margin = margin(t = -5, b = 0, 
                                                   r = 1, l = -8, "mm"),
                              legend.position="none"),
          fig_hp[[5]] + theme(plot.margin = margin(t = -5, b = 0, 
                                                   r = 1, l = -8, "mm"),
                              legend.position="none"),
          # fig_hp[[6]] + theme(plot.margin = margin(t = -5, b = 0, 
          #                                          r = 1, l = -12, "mm"),
          #                     legend.position="none"), 
          nrow = 1)
fig_b2 <- plot_grid(fig_b1, legend, rel_widths = c(0.945, 0.055), 
                    nrow = 1)
hp_vcn <- plot_grid(Fig_4, 
                    fig_vcn + 
                      theme(plot.margin = margin(t = -8, b = 0, 
                                                 r = 1, l = 1, "mm")),
                    fig_b2 + 
                      theme(plot.margin = margin(t = 5, b = 0, 
                                                 r = 1, l = 1, "mm")),
                    rel_heights = c(2, 1, 1.5),
                    rel_widths = 1, nrow = 3,
                    labels = c("a", "b", "c"), 
                    label_y = c(1, 1.2, 1.15))
hp_vcn
```

### Genes in each category

DS genes in each category are sorted by their logFC.
```{r}
gene_list <- lapply(seq_along(list_df),
                    FUN = function(i) {
                      x_df <- list_df[[i]]
                      x_node <- node_list[[i]]
                      
                      v_df <- x_df$variable
                      v_gene <- df_gene %>%
                        filter(gene_sid %in% v_df) %>%
                        mutate_if(is.factor, as.character)
                      
                      out %>%
                        filter(node == x_node) %>%
                        filter(feature %in% v_gene$gene_id) %>%
                        mutate(gene_sid = sub(pattern = ".*?\\.", replacement = "", feature)) %>%
                        arrange(logFC) %>%
                        select(feature) %>% unlist() %>% as.character()
                    })

names(gene_list) <- names(list_df) <- names(node_list)
lapply(gene_list, head)
lapply(gene_list, length)

```



## nodes detected in each category
```{r, fig.width=8, fig.height=8}
# signal nodes (node) -  genes (feature)
df_signal <- topNodes(object = best_0.05, n = Inf, p_value = 0.05)

df_report <- lapply(seq_along(gene_list), FUN = function(x) {
  message(x)
  # frequences of nodes detected in each category
  df.x <- df_signal %>% 
    filter(feature %in% gene_list[[x]])
  xt <- table(df.x$node)
  nodeE <- unlist(node_list, use.names = FALSE)[x]
  data.frame(category = names(gene_list)[x],
             node = as.numeric(names(xt)),
             freq = as.vector(xt),
             prop = as.vector(xt)/length(gene_list[[x]]),
             node_true = nodeE) %>%
    arrange(-prop)
  })

# category : category
# node : detected nodes
# freq: how many genes in each category are detected in each node
# prop: freq/length(gene) in each category
# node_true: the branch node of cell-type specific branch
df_report <- do.call(rbind, df_report)
df_report2 <- df_report %>%
  group_by(category) %>%
  filter(prop > 0.5) %>%
  left_join(Fig_0$data) %>%
  select(category, node, freq, prop, node_true,x, y)
df_report2


# this is avoid overlap in texts
Fig_4u <- Fig_4 +
  new_scale_fill() +
  new_scale_color() +
  geom_point(data = df_report2, aes(x, y, color = category), 
              size = 3, 
             inherit.aes = FALSE, show.legend = FALSE)  +
   geom_text(data = df_report2, 
               aes(x = x, y = y, label = category)) +
guides(fill = guide_legend(order = 2, ncol = 2)) +
  scale_fill_manual(values = range_col) +
  scale_color_manual(values = range_col)


hp_vcn <- plot_grid(Fig_4u, 
                    fig_vcn + 
                      theme(plot.margin = margin(t = -3, b = 0, 
                                                 r = 1, l = 1, "mm")),
                    fig_b2 + 
                      theme(plot.margin = margin(t = 2, b = 0, 
                                                 r = 1, l = -5, "mm")),
                    rel_heights = c(1.8, 1, 1.4),
                    rel_widths = 1, nrow = 3,
                    labels = c("a", "b", "c"), 
                    label_y = c(1, 1.15, 1.05))
hp_vcn
ggsave(filename = "output/LPS.eps", hp_vcn, width = 8, height = 8, units = "in")
```
